graphql-alt: relaxing consistency
## Description

In an earlier change, (along with the introduction of the `Scope`
abstraction), helpers were added for handling `checkpoint_viewed_at`
within cursors, exactly like how they worked in OG GraphQL.

In OG GraphQL, almost all cursors include a `checkpoint_viewed_at`,
which fixed the state of the chain to the point in time at which the
pagination started.

However, we identified that in some cases, this pattern is not
desirable, particularly for streaming access patterns (polling or
subscriptions) that need to keep up (continue returning fresh results)
as the chain progresses. The only paginated queries that must remain
consistent are related to the live object set (filtering objects by
type, or by owner and type).

This change removes the `checkpoint_viewed_at` handling for
`TransactionEffects.objectChanges`, and in general from now on queries
that are not related to the live object set will not include a
`checkpoint_viewed_at` component to their cursor, this means that
queries that could be streamed, can.

The previous behaviour can always be replicated using the
`Checkpoint.query` field.

## Test plan

Existing tests:

```
sui$ cargo nextest run -p sui-indexer-alt-e2e-tests
```