[sui-adapter] Add basic package cache (#22099)
## Description 

Now that we have the `PackageStore` abstraction in the adapter layer,
this adds in a very simple caching layer for `MovePackage`s. Currently
the cache size is bounded at 200 packages, after that we drop the cache
and reload (we could also go for an LRU if we wanted, but I think that
will just be more complexity than is most likely needed here). The cache
is only held for the lifetime of the transaction.

This also does a bit of reorganization around the data store, and
related "store" functionality into its own directory under the
programmable transactions.

Next up: will work on refactoring `LinkageView`

## Test plan 

CI

Co-authored-by: Timothy Zakian <timothyzakian@mac.lan>